Finally made the app to work before razorpay

+++ controllers/expensecontrollers.js
@@ -2,37 +2,44 @@ const Expense = require('../models/Expense');
 
 const addExpense = async (req,res,next)=>{
     try{
-        //console.log('hi')
-        const amount = req.body.amount;
-        const description = req.body.description;
-        const category = req.body.category;
+        const {amount,description,category} = req.body;
+        if(amount === undefined || amount.length === 0){
+            return res.status(400).json({message:'Parameters are missing',success:false});
+        }
 
-        const data = await Expense.create({amount:amount,description:description,category:category}); 
-        res.status(201).json({newExpenseAdded:data});
+        const data = await Expense.create({amount, description, category, userId: req.user.id}); 
+       res.status(201).json({newExpenseAdded:data,success:true});
     } catch(error){
         console.log(JSON.stringify(error));
-        res.status(500).json({error:error});
+        res.status(500).json({message:error,success:false});
     }
 };
 
+
 const getExpense = async (req,res,next)=>{
     try{
-        const expenses = await Expense.findAll({where:{userId:req.user.id}});
-        res.status(200).json({allExpenses:expenses});
+        const expenses = await Expense.findAll({ where: { userId:req.user.id }});
+        res.status(200).json({allExpenses:expenses,success:true});
     } catch(error){
         console.log(JSON.stringify(error));
-        res.status(500).json({error:error});
+        res.status(500).json({message:error,success:false});
     }
 };
 
 const deleteExpense = async (req,res,next)=>{
     try{
         const expenseID = req.params.id;
-        await Expense.destroy({where:{id:expenseID}});
-        res.sendStatus(200);
+        if(expenseID === undefined || expenseID.length === 0){
+            return res.status(400).json({message:'Expense ID is missing',success:false});
+        }
+        const noofrows = await Expense.destroy({where:{id:expenseID,userId:req.user.id}});
+        if(noofrows === 0){
+            return res.status(404).json({message:'expense doesnot belongs to user'});
+        }
+        res.status(200).json({message:'Deleted Successfully',success:true});
     } catch(error){
         console.log(JSON.stringify(error));
-        res.status(500).json({error:error});
+        res.status(500).json({message:error,success:false});
     }
 };
